Orders Initial

+++ src/Services/Orders/Orders.Domain/Events/OrderItemAddedEvent.cs
@@ -8,10 +8,10 @@ namespace Orders.Domain.Events
         public Guid OrderId { get; private set; }
         public Guid ProductId { get; private set; }
         public string ProductName { get; private set; }
-        public double UnitPrice { get; private set; }
+        public decimal UnitPrice { get; private set; }
         public int Quantity { get; private set; }
 
-        public OrderItemAddedEvent(Guid orderId, Guid productId, string productName, double unitPrice, int quantity)
+        public OrderItemAddedEvent(Guid orderId, Guid productId, string productName, decimal unitPrice, int quantity)
         {
             OrderId = orderId;
             ProductId = productId;
